Moved around code to fix circular import issues.
